[ADD] PWA capabilities based on @angular/pwa and @angular/service-worker

+++ src/app/services/auth.service.ts
@@ -7,20 +7,25 @@ import { UserService } from './user.service';
 
 import Fingerprint2 from '@fingerprintjs/fingerprintjs';
 import { Router } from '@angular/router';
+import { StorageService } from './storage.service';
 
 @Injectable()
 export class AuthService {
     private authURL = environment.api + '/auth/';
+    private domain = environment.domain == 'localhost' ? environment.domain : '.' + environment.domain;
     private fingerprint: string;
     private isLoggedIn: boolean = false;
-    
-    constructor(private httpClient: HttpClient, private cookieService: CookieService, private userService: UserService, private router: Router ) {
+    private cookieTTL: any = 0;
+
+    constructor(private httpClient: HttpClient, 
+                private cookieService: CookieService, 
+                private userService: UserService, 
+                private storageService: StorageService,
+                private router: Router ) {
         this.init();
     }
 
     private init(){
-        const user = JSON.parse(sessionStorage.getItem('user'));
-
         Fingerprint2.getPromise().then( (components) => {
             let values = components.map( c => { return c.value })
             let fingerprint = Fingerprint2.x64hash128(values.join(''), 31);         
@@ -28,12 +33,16 @@ export class AuthService {
         })
 
         if(this.session && this.session !== ''){
-            this.isLoggedIn = true;
+            this.isLoggedIn = true; 
         }
+    }
 
-        if(user){
-            this.userService.set(user);
-        }
+    private clear(){
+        this.storageService.clearLocalStorage();     
+        this.cookieService.deleteAll('/', this.domain);
+        this.router.navigate(['/']);                             
+        this.userService.set(null);
+        this.isLoggedIn = false;  
     }
 
     get loggedIn(): boolean{
@@ -46,18 +55,19 @@ export class AuthService {
 
 
 
-    login(user: string, password: string, remember: boolean){
+    login(user: string, password: string, remember: boolean){        
+        if(remember){
+            this.cookieTTL = new Date();
+            this.cookieTTL.setFullYear(this.cookieTTL.getFullYear() + 1);
+        }
+        
         return this.httpClient.post(this.authURL, { user: user, password: password, device: this.fingerprint }).pipe(tap(
             res => {
-                let date = new Date();
-                
-                date.setFullYear(date.getFullYear() + 1);
-                                
                 this.userService.set(res['user']);
-                this.cookieService.set("uid", res['user']['_id'], 0, '/');
-                this.cookieService.set("token", res['token'], 0, '/');
-                this.cookieService.set("fingerprint", this.fingerprint, remember ? date : 0, '/');
-                this.cookieService.set("session", res['session'], remember ? date : 0, '/');
+                this.cookieService.set("token", res['token'], 0, '/', this.domain);
+                this.cookieService.set("uid", res['user']['_id'], this.cookieTTL, '/', this.domain);
+                this.cookieService.set("fingerprint", this.fingerprint, this.cookieTTL, '/', this.domain);
+                this.cookieService.set("session", res['session'], this.cookieTTL, '/', this.domain);
 
                 this.isLoggedIn = true;
                 this.router.navigate(['/home']);
@@ -72,22 +82,23 @@ export class AuthService {
 
     logout(device:boolean = true){
         const fingerprint = this.cookieService.get("fingerprint");
+        const token = this.cookieService.get("token");
+        const session = this.cookieService.get("session");
+        
+        this.clear();
 
+        if(!session && !token){
+            return;
+        }
+        
         return this.httpClient.delete(this.authURL, { params: device ? { device: fingerprint } : {} }).subscribe(
-            res => {                
+            res => {      
                 return res["success"];
             },
             err => {
                 return err
             }            
-        ).add(() => {
-            this.cookieService.delete('session');
-            this.cookieService.delete('token');
-            this.cookieService.delete('uid');             
-            this.router.navigate(['/']);                             
-            this.userService.set(null);
-            this.isLoggedIn = false;   
-        });
+        )
     }
 
     refresh(){
@@ -95,22 +106,15 @@ export class AuthService {
 
         return this.httpClient.patch(this.authURL, { user:session.uid, token:session.session, device: session.fingerprint }).pipe(tap(
             res => {
-                let date = new Date();
-               
-                date.setFullYear(date.getFullYear() + 1);
-
                 this.userService.set(res['user']);
-                this.cookieService.set("token", res['token'], 0, '/');
-                this.cookieService.set("session", res['session']);
-
+                this.cookieService.set("token", res['token'], 0, '/', this.domain);
+                this.cookieService.set("session", res['session'], this.cookieTTL, '/', this.domain);
                 this.isLoggedIn = true;
                 
                 return res;
             },
             err => {
-                this.isLoggedIn = false;
-                this.cookieService.deleteAll();
-                this.userService.set(null);
+                this.clear();
                 return err
             }
         ));
